+++ src/args.js
@@ -1,6 +1,0 @@
-/*
- * natron
- */
-import minimist from "minimist";
-
-export default minimist(process.argv.slice(2));