+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # CHANGELOG
 ## master
+- Use `natron-logging`
+
 ## v0.1.0 (2015-09-14)
 - **Initial release**